# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Wakarimi Music Files API",
        "contact": {
            "name": "Dmitry Kolesnikov (Zalimannard)",
            "email": "zalimannard@mail.ru"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.4"
    },
    "host": "localhost:8022",
    "basePath": "/api",
    "paths": {
...
        "/songs/sha256/{sha256}": {
            "get": {
                "description": "Retrieves a list of songs that have the specified SHA256 hash.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Search songs by SHA256 hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SHA256 hash",
                        "name": "sha256",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/song_handler.searchBySha256Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
        "/songs/{songId}": {
            "get": {
                "description": "Retrieves a single song by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Retrieve a song by its ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Song ID",
                        "name": "songId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/song_handler.getSongResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responses.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "song_handler.getSongResponse": {
            "type": "object",
            "properties": {
                "bitrateKbps": {
                    "description": "Bitrate in kilobits per second",
                    "type": "integer"
                },
                "channelsN": {
                    "description": "Number of audio channels",
                    "type": "integer"
                },
                "dirId": {
                    "description": "Directory identifier where the song resides",
                    "type": "integer"
                },
                "durationMs": {
                    "description": "Duration of the song in milliseconds",
                    "type": "integer"
                },
                "extension": {
                    "description": "File extension of the song",
                    "type": "string"
                },
                "filename": {
                    "description": "Filename of the song",
                    "type": "string"
                },
                "lastContentUpdate": {
                    "description": "Time of the last update to the song's content",
                    "type": "string"
                },
                "sampleRateHz": {
                    "description": "Sample rate in hertz",
                    "type": "integer"
                },
                "sha256": {
                    "description": "SHA-256 hash of the file",
                    "type": "string"
                },
                "sizeByte": {
                    "description": "File size in bytes",
                    "type": "integer"
                },
                "songId": {
                    "description": "Unique identifier for the song",
                    "type": "integer"
                }
            }
        },
...
        "song_handler.searchBySha256Response": {
            "type": "object",
            "properties": {
                "songs": {
                    "description": "Array of songs that match the search query.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/song_handler.searchBySha256ResponseItem"
                    }
                }
            }
        },
        "song_handler.searchBySha256ResponseItem": {
            "type": "object",
            "properties": {
                "bitrateKbps": {
                    "description": "Bitrate of the song in Kbps.",
                    "type": "integer"
                },
                "channelsN": {
                    "description": "Number of channels in the song.",
                    "type": "integer"
                },
                "dirId": {
                    "description": "Directory ID where the song is located.",
                    "type": "integer"
                },
                "durationMs": {
                    "description": "Duration of the song in milliseconds.",
                    "type": "integer"
                },
                "extension": {
                    "description": "File extension of the song.",
                    "type": "string"
                },
                "filename": {
                    "description": "Filename of the song.",
                    "type": "string"
                },
                "lastContentUpdate": {
                    "description": "Timestamp of the last content update.",
                    "type": "string"
                },
                "sampleRateHz": {
                    "description": "Sample rate of the song in Hz.",
                    "type": "integer"
                },
                "sha256": {
                    "description": "SHA-256 hash of the song.",
                    "type": "string"
                },
                "sizeByte": {
                    "description": "File size of the song in bytes.",
                    "type": "integer"
                },
                "songId": {
                    "description": "Unique identifier for the song.",
                    "type": "integer"
                }
            }
        }
    }
}